+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.3.0] - 2021-01-26
+## [2.3.0] - 2021-01-27
 ### Added
 - Integrate API apps into the CLI. Support API message transmission, reception,
   and the demo-rx app directly from `blocksat-cli` commands (`api send`, `api
@@ -48,8 +48,10 @@
 - Alternative option on the SDR app to define the logical channel that the
   receiver tunes to. The new option is named `--chanel` and is mutually
   exclusive with the pre-existing `--modcod` option.
-- Support for the reception of Lightning gossip snapshots transported via the
-  satellite API.
+- Add option `--gossip` on the API listener and demo-rx apps to support the
+  reception of Lightning gossip snapshots transported via the satellite API.
+- Add option `--btc-src` on the API listener and demo-rx apps to support the
+  reception of Bitcoin source code files sent via the satellite API.
 - Support specification of USB DVB-S2 adapter by model name on `blocksat-cli
   usb` command.
 - Support for non-interactive dvbnet interface removal.